render-github-files/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dateutil==2.9.0.post0
numpy==2.1.3
pandas==2.2.3
psycopg[binary]==3.*
